Use Express
In general, these days I would prefer Fastify for my projects.

It's a [relatively new][1] framework, but already proved to be
production-ready. It combines the best of three the most popular
frameworks (Express, Hapi and Restify).

[Performance benchmarks][2] demonstrate that Fastify tends to be the
fastest, but in real life it's [not only about performance][3] when it
comes to such kind of choices. First of all, performance benchmarks are
[doubtful][4] and a subject to debate. Secondly, we choose frameworks
based not only on their ability to process bare requests, but also on
features they have (or don't), expertise we have or don't, the way they
match our priorities and needs, community support, maintainability, etc.
Multiple factors in other words ;)

Anyways, since you guys use mostly Express I decided to go with it as
well for this assignment :)

[1]:
https://thenewstack.io/introducing-fastify-speedy-node-js-web-framework/
[2]: https://github.com/fastify/fastify#benchmarks
[3]: https://twitter.com/tjholowaychuk/status/1098589765463609345
[4]: https://hueniverse.com/performance-at-rest-75bb8fff143